@@ -1,7 +1,7 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -10,11 +10,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\DoePaAdminDataModel\DoePaAdminDataModel.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="DPApp\SQLCommands\" />
-    <Folder Include="Resources\" />
   </ItemGroup>
 
   <ItemGroup>
