@@ -1,16 +1,33 @@
-﻿using DoePaAdminDataModel.Kostenrechnung;
+﻿using DoePaAdminApp.Models;
+using DoePaAdminDataAdapter.DPApp.Model;
 using System;
 using System.Collections.Generic;
-using System.Text;
+using System.Data;
+using Microsoft.Extensions.Options;
+using DoePaAdminDataAdapter.DPApp;
+using System.Threading.Tasks;
 
 namespace DoePaAdminApp.Services
 {
     class DPAppService : IDPAppService
     {
 
-        public ICollection<Ausgangsrechnung>  GetAusgangsrechnungen()
+        public string DPAppConnectionString { get; set; }
+
+        public DPAppService(IOptions<DPAppConnectionSettings> dpAppSettings)
+        {
+            DPAppConnectionString = dpAppSettings.Value.ConnectionString;
+        }
+
+        public ICollection<OutgoingInvoice> GetAusgangsrechnungen()
         {
             throw new NotImplementedException();
+        }
+
+        public async Task<DataTable> GetCostCentersAsync()
+        {
+            MasterdataDAL dal = new(DPAppConnectionString);
+            return await dal.ReadCostCenterAsync();
         }
 
     }
