@@ -1,14 +1,16 @@
-﻿using DoePaAdminDataModel.Kostenrechnung;
-using System;
+﻿using DoePaAdminDataAdapter.DPApp.Model;
 using System.Collections.Generic;
-using System.Text;
+using System.Data;
+using System.Threading.Tasks;
 
 namespace DoePaAdminApp.Services
 {
     public interface IDPAppService
     {
 
-        ICollection<Ausgangsrechnung> GetAusgangsrechnungen();
+        ICollection<OutgoingInvoice> GetAusgangsrechnungen();
+
+        Task<DataTable> GetCostCentersAsync();
 
     }
 }
